OSX changes, StepVis, ColorMap
Changed event->rect() to rect() so things draw correctly in OSX 10.8.
First pass of StepVis, has not been confirmed to work
ColorMap class is also new, has not been confirmed to work.

+++ viswidget.cpp
@@ -14,7 +14,7 @@ VisWidget::VisWidget(QWidget *parent)
     visProcessed = false;
 }
 
-void VisWidget::setSteps(int start, int stop)
+void VisWidget::setSteps(float start, float stop)
 {
 
 }
